Got most of the source code compiling on Mac but need other tests on windows to confirm also I can bring back the mesh support now that rendering api isnt a thing

+++ Engine/Util/Include/VHash.h
@@ -5,7 +5,7 @@
 #include "UtilAPI.h"
 
 //the typedef for a hash
-typedef size_t vhash_int;
+typedef uint32 vhash_int;
 
 /**
 * This class is used to optimize the comparison between 2 strings
